Just specify work item size instead of batch * num
Batch size and batches per work item was an artifact of when
the system was designed for the very specific purpose of evaluating
neural networks on video. Now that it is more general, each evaluator
should handle the batching within itself while work item size should
be based on load balancing and scheduling overhead.